More Travis fixes.

+++ src/c4/test/tstInvert_Comm_Map.cc
@@ -37,15 +37,15 @@ void test2(rtt_c4::ParallelUnitTest &ut) {
   invert_comm_map(to_nodes, from_nodes);
 
   if (node == 0) {
-    if (from_nodes.size() != 0)
-      FAILMSG("Incorrect map size on node 0.");
+      if (static_cast<int>(from_nodes.size()) != 0)
+          FAILMSG("Incorrect map size on node 0.");
   }
 
   if (node == 1) {
-    if (from_nodes.size() != 1)
-      FAILMSG("Incorrect size of map on node 1.");
-    if (from_nodes[0] != 0)
-      FAILMSG("Incorrect map contents on node 1.");
+      if (static_cast<int>(from_nodes.size()) != 1)
+          FAILMSG("Incorrect size of map on node 1.");
+      if (from_nodes[0] != 0)
+          FAILMSG("Incorrect map contents on node 1.");
   }
 
   if (ut.numFails == 0)
@@ -62,18 +62,18 @@ void test4(rtt_c4::ParallelUnitTest &ut) {
   std::vector<int> to_nodes;
 
   if (node == 0) {
-    to_nodes.push_back(1);
-    to_nodes.push_back(2);
-    to_nodes.push_back(3);
+      to_nodes.push_back(1);
+      to_nodes.push_back(2);
+      to_nodes.push_back(3);
   }
   if (node == 1) {
-    to_nodes.push_back(0);
+      to_nodes.push_back(0);
   }
   if (node == 2) {
-    to_nodes.push_back(0);
+      to_nodes.push_back(0);
   }
   if (node == 3) {
-    to_nodes.push_back(0);
+      to_nodes.push_back(0);
   }
 
   std::vector<int> from_nodes(0);
@@ -81,17 +81,17 @@ void test4(rtt_c4::ParallelUnitTest &ut) {
   invert_comm_map(to_nodes, from_nodes);
 
   if (node == 0) {
-    if (from_nodes.size() != 3)
-      FAILMSG("Incorrect map size on node 0");
-    for (int i = 0; i < 3; ++i) {
-      if (from_nodes[i] != i+1)
-        FAILMSG("Incorrent map contents on node 0");
+      if (static_cast<int>(from_nodes.size()) != 3)
+          FAILMSG("Incorrect map size on node 0");
+      for (int i = 0; i < 3; ++i) {
+          if (from_nodes[i] != i+1)
+              FAILMSG("Incorrent map contents on node 0");
     }
   } else {
-    if (from_nodes.size() != 1)
-      FAILMSG("Incorrect map size.");
-    if (from_nodes[0] != 0)
-      FAILMSG("Incorrect map contents.");
+      if (static_cast<int>(from_nodes.size()) != 1)
+          FAILMSG("Incorrect map size.");
+      if (from_nodes[0] != 0)
+          FAILMSG("Incorrect map contents.");
   }
 
   if (ut.numFails == 0)
@@ -114,7 +114,7 @@ void test_n_to_n(rtt_c4::ParallelUnitTest &ut) {
   std::vector<int> from_nodes;
   invert_comm_map(to_nodes, from_nodes);
 
-  if (from_nodes.size() != nodes)
+  if (static_cast<int>(from_nodes.size()) != nodes)
       FAILMSG("Incorrect from_nodes size.");
 
   for (int i = 0; i < nodes; ++i) {
@@ -142,7 +142,7 @@ void test_cyclic(rtt_c4::ParallelUnitTest &ut) {
   std::vector<int> from_nodes;
   invert_comm_map(to_nodes, from_nodes);
 
-  if (from_nodes.size() != 1)
+  if (static_cast<int>(from_nodes.size()) != 1)
     FAILMSG("Incorrect map size.");
   if (from_nodes[0] != (node + nodes - 1) % nodes)
     FAILMSG("Incorrect map contents in cyclc test.");
@@ -163,7 +163,7 @@ void test_empty(rtt_c4::ParallelUnitTest &ut) {
 
   invert_comm_map(to_nodes, from_nodes);
 
-  if (from_nodes.size() != 0)
+  if (static_cast<int>(from_nodes.size()) != 0)
     FAILMSG("Incorrect map size in empty test.");
 
   if (ut.numFails == 0)
